backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.135.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-deepseek==1.0.1
langchain-experimental==0.4.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.5.0
llama-index-embeddings-dashscope>=0.3.0
dashscope>=1.22.0
llama-index-instrumentation>=0.5.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==82.0.1
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
mem0ai>=0.1.0
qdrant-client>=1.9.0
